backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.6.1

# Google Cloud and AI dependencies
google-cloud-aiplatform==1.42.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-genai==0.3.2
vertexai==1.42.1

# Audio processing dependencies
google-cloud-speech==2.28.0
pydub==0.25.1
ffmpeg-python==0.2.0

# Google ADK (Agent Development Kit)
google-adk==0.1.0

# Environment and configuration
python-dotenv==1.0.1

# HTTP requests
requests==2.31.0

# File handling and utilities
python-multipart==0.0.9

# CORS middleware
fastapi-cors==0.0.6

# LlamaIndex for RAG (if needed)
llama-index==0.10.8

# Additional utilities
typing-extensions==4.9.0 